blog_generation/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.3
langchain-openai==0.0.3
langchain-anthropic==0.1.1
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6
jinja2==3.1.3
beautifulsoup4==4.12.3
lxml==5.1.0
pytest==7.4.4
pytest-asyncio==0.21.1